+++ client/web/src/hooks/index.ts
@@ -1,3 +1,4 @@
-export { useRoutesMatch } from './useRoutesMatch'
 export { useHandleSubmitFeedback } from './useHandleSubmitFeedback'
+export { useRoutesMatch } from './useRoutesMatch'
 export { useURLSyncedState } from './useUrlSyncedState'
+export { useScrollManager } from './useScrollManager'
