+++ client/web/src/hooks/useScrollManager/index.ts
@@ -1,0 +1,1 @@
+export { useScrollManager } from './useScrollManager'
